+++ app/src/main/java/com/myth/mama/MainActivity.kt
@@ -3,7 +3,7 @@ package com.myth.mama
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.myth.arch.mvvm2.MythView
+import com.myth.arch.mvvm3.MythView
 import com.myth.arch.mvvm3.toast
 import kotlinx.android.synthetic.main.activity_main.*
