--- myth-arch-mvvm-lib/src/main/java/com/myth/arch/mvvm2/MythActivity.kt
@@ -1,9 +1,0 @@
-package com.myth.arch.mvvm2
-
-import androidx.appcompat.app.AppCompatActivity
-
-/**
- * Reserve class
- */
-open class MythActivity : AppCompatActivity(), MythView {
-}
