--- myth-arch-mvvm-lib/src/main/java/com/myth/arch/mvvm2/MythFragment.kt
@@ -1,9 +1,0 @@
-package com.myth.arch.mvvm2
-
-import androidx.fragment.app.Fragment
-
-/**
- * Reserve class
- */
-class MythFragment : Fragment(), MythView {
-}
