--- myth-arch-mvvm-lib/src/main/java/com/myth/arch/mvvm2/log/MythLogger.kt
@@ -1,4 +1,0 @@
-package com.myth.arch.mvvm2.log
-
-object MythLogger {
-}
