+++ myth-arch-mvvm-lib/src/main/java/com/myth/arch/mvvm3/MythType.kt
@@ -1,6 +1,5 @@
 package com.myth.arch.mvvm3
 
 import androidx.lifecycle.LiveData
-import com.myth.arch.mvvm2.MythView
 
 typealias MythViewModelExt<T> = (MythView, LiveData<T>) -> Unit
